Moved non-file I/O spec conversion functions into sdk-codegen
- Minor cleanup for testUtils.ts
- Spec file I/O is still in sdk-codegen-scripts
- Fixed some miscellaneous test issues
