+++ packages/sdk-codegen/src/index.ts
@@ -35,3 +35,4 @@ export * from './csharp.gen'
 export * from './codeGenerators'
 export * from './exampleInfo'
 export * from './specLinter'
+export * from './specConverter'
