+++ packages/sdk/test/methods.spec.ts
@@ -372,10 +372,13 @@ describe('LookerNodeSDK', () => {
 
     it('search_looks fields and title', async () => {
       const sdk = new LookerSDK(session)
+      const looks = await sdk.ok(sdk.all_looks('id,title'))
+      expect(looks).not.toHaveLength(0)
+      const expected = looks[0]
       const actual = await sdk.ok(
         sdk.search_looks({
           fields: 'id,title',
-          title: 'An SDK%',
+          title: expected.title,
         })
       )
       expect(actual).toBeDefined()
@@ -383,7 +386,7 @@ describe('LookerNodeSDK', () => {
       const look = actual[0]
       expect(look.id).toBeDefined()
       expect(look.title).toBeDefined()
-      expect(look.title).toContain('SDK')
+      expect(look.title).toEqual(expected.title)
       expect(look.description).not.toBeDefined()
       await sdk.authSession.logout()
       expect(sdk.authSession.isAuthenticated()).toBeFalsy()
